added error handling for BadCredentials and AccessDenied exceptions and moved final strings to apiConfig
