+++ backend/src/main/java/com/intive/shopme/config/AppConfig.java
@@ -47,6 +47,11 @@ public final class AppConfig {
     public static final String EMAIL_CLAIM_NAME = "email";
     public static final String SCOPES_CLAIM_NAME = "scopes";
 
-    private AppConfig() {
+    public static final String CONSTRAINTS_JSON_KEY = "errors";
+    public static final String VALIDATION_DESCRIPTION_JSON_KEY = "message";
+    public static final String ERROR_ID_JSON_KEY = "id";
+    public static final String ACCESS_DENIED = "Access denied";
+
+    public AppConfig() {
     }
 }
